feat: add update-live command
